feat(controller): enhance arm action to reset origin before arming
This commit modifies the behavior of the `arm` action within the `InterfaceActions` structure in `controller.cpp`. The change introduces a call to `origin_reset->resetOrigin()` prior to setting the arm state on the PX4.

The modification aims to ensure that the origin is reset each time the arm command is issued, which is critical for ensuring accurate positioning and navigation of the UAV after arming. This adjustment contributes to improved handling and reliability of the UAV's response to control commands.

BREAKING CHANGE: The behavior of the arm action has changed; it now includes an origin reset, which may affect previous interactions that relied on the prior state of the system when arming.